[GH#69] Add Ability to Center Password Input Text
Add a new possible `"center"` value to the `password-alignment` option,
specifying that the text in the password input should be centered,
instead of left- or right-aligned.

Closes GH#69

+++ src/ui.c
@@ -226,7 +226,7 @@ static void create_and_attach_password_field(Config *config, UI *ui)
         gtk_entry_set_invisible_char(GTK_ENTRY(ui->password_input), *config->password_char);
     }
     gtk_entry_set_alignment(GTK_ENTRY(ui->password_input),
-                            (gfloat) config->password_alignment);
+                            config->password_alignment);
     // TODO: The width is usually a little shorter than we specify. Is there a
     // way to force this exact character width?
     // Maybe use 2 GtkBoxes instead of a GtkGrid?
